removed gradle-wrapper and other files, renamed the package of language server, deleted generated comments
