+++ org.eclipse.buildship.gradleprop.languageserver/build.gradle.kts
@@ -1,28 +1,20 @@
 plugins {
-    // Apply the application plugin to add support for building a CLI application in Java.
     application
     id("com.github.johnrengelman.shadow") version "7.0.0"
 }
 
 repositories {
-    // Use Maven Central for resolving dependencies.
     mavenCentral()
 }
 
 dependencies {
-    // Use JUnit Jupiter for testing.
-    testImplementation("org.junit.jupiter:junit-jupiter:5.9.1")
-
-    implementation("com.fasterxml.jackson.core:jackson-databind:2.0.1")
-
-    //    implementation
+    testImplementation("org.junit.jupiter:junit-jupiter:5.9.2")
+    implementation("com.fasterxml.jackson.core:jackson-databind:2.15.1")
     implementation("org.eclipse.lsp4j:org.eclipse.lsp4j:0.21.0")
     implementation("org.eclipse.lsp4j:org.eclipse.lsp4j.jsonrpc:0.21.0")
-    // This dependency is used by the application.
     implementation("com.google.guava:guava:31.1-jre")
 }
 
-// Apply a specific Java toolchain to ease working on different environments.
 java {
     toolchain {
         languageVersion.set(JavaLanguageVersion.of(11))
@@ -36,12 +28,10 @@ tasks.withType(Jar::class) {
 }
 
 application {
-    // Define the main class for the application.
-    mainClass.set("server.GradlePropertiesLanguageServerLauncher")
+    mainClass.set("org.eclipse.buildship.gradleprop.ls.GradlePropertiesLanguageServerLauncher")
 }
 
 tasks.named<Test>("test") {
-    // Use JUnit Platform for unit tests.
     useJUnitPlatform()
 }
