+++ org.eclipse.buildship.gradleprop.languageserver/src/main/java/org/eclipse/buildship/gradleprop/ls/GradlePropertiesLanguageServerLauncher.java
@@ -1,4 +1,4 @@
-package server;
+package org.eclipse.buildship.gradleprop.ls;
 
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
