+++ org.eclipse.buildship.gradleprop.languageserver/src/main/java/org/eclipse/buildship/gradleprop/ls/completion/PropertyModelJson.java
@@ -1,4 +1,4 @@
-package server.completion;
+package org.eclipse.buildship.gradleprop.ls.completion;
 
 import java.util.ArrayList;
